+++ src/main/java/com/skhu/cloud/dto/diff/FileDiffDto.java
@@ -1,0 +1,4 @@
+package com.skhu.cloud.dto.diff;
+
+public class FileDiffDto {
+}
