+++ src/main/java/com/skhu/cloud/dto/diff/FolderDiffDto.java
@@ -1,0 +1,7 @@
+package com.skhu.cloud.dto.diff;
+
+import com.skhu.cloud.dto.FileDto;
+
+public class FolderDiffDto extends FileDto {
+    private Long flag; // flag 만 추가
+}
