+++ src/main/java/com/skhu/cloud/dto/error/CustomException.java
@@ -1,4 +1,4 @@
-package com.skhu.cloud.dto;
+package com.skhu.cloud.dto.error;
 
 public class CustomException extends Exception {
