+++ src/main/java/com/skhu/cloud/dto/error/EnumModel.java
@@ -1,8 +1,6 @@
-package com.skhu.cloud.dto;
+package com.skhu.cloud.dto.error;
 
 public interface EnumModel {
-
     String getKey();
     String getValue();
-
 }
